requirements.txt
moviepy==1.0.3
Pillow
nest_asyncio
google-generativeai
google-genai
typing_extensions
openai
IPython
requests
opencv-python
pyside6